requirements.txt
# Python Django Dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
requests==2.31.0
Pillow==10.1.0
pypdf2==3.0.1
python-docx==1.1.0
